(feat) Built-ins exclusion from path_finder function
