+++ src/includes/minishell.h
@@ -6,7 +6,7 @@
 /*   By: Juliany Bernardo <julberna@student.42sp    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/10 13:44:05 by julberna          #+#    #+#             */
-/*   Updated: 2024/02/08 20:34:08 by Juliany Ber      ###   ########.fr       */
+/*   Updated: 2024/02/08 20:47:56 by Juliany Ber      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -108,6 +108,7 @@ void			replace_variable(t_token **tokens, t_hash **ht);
 void			expand_variables(t_token **tokens, t_hash **ht);
 
 /* PARSER */
+int				not_builtin(char *cmd);
 int				has_other_pipes(t_token *tokens);
 void			remove_quotes(t_token **tokens);
 void			append_wle(t_word *w, t_word **wl);
