+++ src/parser.c
@@ -6,7 +6,7 @@
 /*   By: Juliany Bernardo <julberna@student.42sp    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/17 17:55:58 by iusantos          #+#    #+#             */
-/*   Updated: 2024/02/08 20:33:01 by Juliany Ber      ###   ########.fr       */
+/*   Updated: 2024/02/08 20:39:10 by Juliany Ber      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -75,49 +75,4 @@ t_ast	*parse_cmd(t_token **tokens, t_ast *parent)
 		}
 	}
 	return (cmd_node);
-}
-
-void	get_path(t_ast **ast, t_hash **env_vars)
-{
-	int		i;
-	char	**paths;
-	char	*full_path;
-
-	if (!*ast)
-		return ;
-	if ((*ast)->type == CMD)
-	{
-		full_path = grab_value("PATH", env_vars);
-		paths = ft_split(full_path, ':');
-		find_path(ast, paths);
-		free(full_path);
-		i = -1;
-		while (paths[++i] != NULL)
-			free(paths[i]);
-		free(paths);
-	}
-	get_path(&(*ast)->left, env_vars);
-	get_path(&(*ast)->right, env_vars);
-}
-
-void	find_path(t_ast **ast, char **paths)
-{
-	int		i;
-	char	*try;
-	char	*temp;
-
-	i = 0;
-	while (paths[i])
-	{
-		temp = ft_strjoin(paths[i], "/");
-		try = ft_strjoin(temp, (*ast)->data->word_list->word);
-		free(temp);
-		if (access(try, X_OK) == 0)
-		{
-			(*ast)->data->pathname = try;
-			break ;
-		}
-		free(try);
-		i++;
-	}
 }
