+++ src/parser_utils.c
@@ -6,7 +6,7 @@
 /*   By: Juliany Bernardo <julberna@student.42sp    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/24 20:24:28 by julberna          #+#    #+#             */
-/*   Updated: 2024/02/08 20:31:49 by Juliany Ber      ###   ########.fr       */
+/*   Updated: 2024/02/08 20:57:33 by Juliany Ber      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -60,6 +60,7 @@ void	set_cmd(t_ast **cmd_node, t_ast **parent)
 	(*cmd_node)->data = malloc(sizeof(t_cmd));
 	if ((*cmd_node)->data == NULL)
 		return ;
+	(*cmd_node)->data->pathname = NULL;
 	(*cmd_node)->data->redirects = NULL;
 	(*cmd_node)->data->word_list = NULL;
 }
